Add alternative Navigate implementation
This version looks better since the navigation happens in one go. OnNavigatedTo is only called on the one page you are going to and OnNavigatedFrom only happens on the page you are leaving (rather than every page in between). It probably is possible to make these events happen on the in-between pages if that would be the best behaviour.